+++ packages/patternfly-4/react-integration/demo-app-ts/src/Demos.ts
@@ -21,14 +21,34 @@ export const Demos: DemoInterface[] = [
     componentType: Examples.AvatarDemo
   },
   {
-    id: 'label-demo',
-    name: 'Label Demo',
-    componentType: Examples.LabelDemo
+    id: 'brand-demo',
+    name: 'Brand Demo',
+    componentType: Examples.BrandDemo
   },
   {
     id: 'chipgroup-demo', 
     name: 'ChipGroup Demo',
     componentType: Examples.BadgeChip 
+  },
+  {
+    id: 'empty-state-demo',
+    name: 'Empty State Demo',
+    componentType: Examples.EmptyStateDemo
+  },
+  {
+    id: 'label-demo',
+    name: 'Label Demo',
+    componentType: Examples.LabelDemo
+  },
+  {
+    id: 'login-page-demo',
+    name: 'Login Page Demo',
+    componentType: Examples.LoginPageDemo
+  },
+  {
+    id: 'modal-demo',
+    name: 'Modal Demo',
+    componentType: Examples.ModalDemo
   },
   {
     id: 'nav-demo',
@@ -41,9 +61,19 @@ export const Demos: DemoInterface[] = [
     componentType: Examples.PopoverDemo
   },
   {
+    id: 'radio-demo',
+    name: 'Radio Demo',
+    componentType: Examples.RadioDemo
+  },
+  {
     id: 'tab-demo',
     name: 'Tab Demo',
     componentType: Examples.TabDemo
+  },
+  {
+    id: 'text-demo',
+    name: 'Text Demo',
+    componentType: Examples.TextDemo
   },
   {
     id: 'title-demo',